Only testing, not editing: XML indentation and whitespace
This generally includes the possibility of writing functions for
processing.

I would prefer to cut the dependency of emacs, but the problem is that
the alternatives are not great. Both `xmllint` and HTML Tidy (`tidy`)
might change the linebreaks of the files during indentation and
whitespace checking, and I haven't seen a way of toggling that off.
Since we don't want the script to make changes to the files, just
validate them, that is not a good option.

XML Lint can be used for checking for a valid and well-formed document,
I think. Just not the indentation and whitespace.