Error handling of Emacs call
This involves an inelegant hack of the order or return values as Emacs
returns more info than we would prefer to stderr when called with
--batch (all that goes to the minibuffer goes to stderr). So to avoid
returning errors (and halting commit) when it actually just returns
diagnostic info, all emacs output to the minibuffer and hence stderr is
dumped and a custom err message is returned when the check fails. This
is routed through stdout to the err vars for further processing in the
script.

I'm not proud of that, but as it is difficult to get rid of the Emacs
dependency, this works for now.